Expanded tests, removed login requirement to allow testing
