+++ web_server/tests/views/test_posts_views.py
@@ -1,97 +1,11 @@
 from django.test import TestCase, Client
 from django.urls import reverse, resolve
 
-from users.models import Author
-from posts.models import Post
-
-
 
 class TestPostsView(TestCase):
 
     def setUp(self):
-        self.fixture_authors = [
-            Author(username="test_user_A", email="test_a@a.com", password="pass_a", first_name="A", last_name="A"),
-            Author(username="test_user_B", email="test_b@b.com", password="pass_b", first_name="B", last_name="B"),
-            Author(username="test_user_C", email="test_c@c.com", password="pass_c", first_name="C", last_name="C")
-        ]
-        for author in self.fixture_authors:
-            author.is_active = 1
-            author.save()
-
-        self.fixture_posts_public = [
-            Post(title='A post 1 public', content="A1 content", author=self.fixture_authors[0]),
-            Post(title='A post 2 public', content="A2 content", author=self.fixture_authors[0]),
-            Post(title='A post 3 public', content="A3 content", author=self.fixture_authors[0]),
-
-            Post(title='B post 1 public', content="B1 content", author=self.fixture_authors[1]),
-            Post(title='B post 2 public', content="B2 content", author=self.fixture_authors[1]),
-            Post(title='B post 3 public', content="B3 content", author=self.fixture_authors[1]),
-
-            Post(title='C post 1 public', content="C1 content", author=self.fixture_authors[2]),
-            Post(title='C post 2 public', content="C2 content", author=self.fixture_authors[2]),
-            Post(title='C post 3 public', content="C3 content", author=self.fixture_authors[2]),
-        ]
-        for post in self.fixture_posts_public:
-            post.size = 0
-            post.visibility = Post.PUBLIC
-            post.contentType = "text/plain"
-            post.save()
-
-        self.fixture_posts_other = [
-            Post(author=self.fixture_authors[0], content="A foaf content", visibility=Post.FOAF, title='A post foaf'),
-            Post(author=self.fixture_authors[0], content="A friend content", visibility=Post.FRIENDS, title='A post friends'),
-            Post(author=self.fixture_authors[0], content="A private content", visibility=Post.PRIVATE, title='A post private'),
-            Post(author=self.fixture_authors[0], content="A server content", visibility=Post.SERVERONLY, title='A post server'),
-
-            Post(author=self.fixture_authors[1], content="B foaf content", visibility=Post.FOAF, title='B post foaf'),
-            Post(author=self.fixture_authors[1], content="B friend content", visibility=Post.FRIENDS, title='B post friends'),
-            Post(author=self.fixture_authors[1], content="B private content", visibility=Post.PRIVATE, title='B post private'),
-            Post(author=self.fixture_authors[1], content="B server content", visibility=Post.SERVERONLY, title='B post server'),
-
-            Post(author=self.fixture_authors[2], content="C foaf content", visibility=Post.FOAF, title='C post foaf'),
-            Post(author=self.fixture_authors[2], content="C friend content", visibility=Post.FRIENDS, title='C post friends'),
-            Post(author=self.fixture_authors[2], content="C private content", visibility=Post.PRIVATE, title='C post private'),
-            Post(author=self.fixture_authors[2], content="C server content", visibility=Post.SERVERONLY, title='C post server'),
-        ]
-        for post in self.fixture_posts_other:
-            post.size = 0
-            post.contentType = 'text/plain'
-            post.save()
-
-        self.endpoints = {
-            "service/posts": reverse('post_index'),
-            "service/posts/{post_id}": reverse("post", )
-        }
-
-        self.client = Client(ACCEPTS="application/json")
-
-    def tearDown(self):
-        for post in self.fixture_posts_other:
-            post.delete()
-        for post in self.fixture_posts_public:
-            post.delete()
-        for author in self.fixture_authors:
-            author.delete()
+        self.client = Client
 
     def test_sample(self):
         assert(1 == 1)
-
-    # Needed tests
-    # service/posts
-    # service/posts/{post_id}
-
-    ################
-    # Tests for service/posts
-    ################
-
-    def test_posts_get_all_public_posts(self):
-        """
-        Test if service/posts will return all public posts
-        :return:
-        """
-        endpoint = self.endpoints['service/posts']
-
-        response = self.client.get(endpoint)
-        print(response)
-
-
